// maplibre/src/client/models/ConnectionEnergyKindEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum ConnectionEnergyKindEnum {
  CONSUMPTION = 'consumption',
  PRODUCTION = 'production',
  CONSUMPTION_PRODUCTION = 'consumptionProduction',
  OTHER = 'other',
}